Separate out netcanvasFile functions with interdependence
