+++ src/utils/netcanvasFile/netcanvasFile.js
@@ -1,4 +1,3 @@
-import { remote } from 'electron';
 import fse from 'fs-extra';
 import log from 'electron-log';
 import path from 'path';
@@ -9,22 +8,21 @@ import canUpgrade from '@app/protocol-validation/migrations/canUpgrade';
 import migrateProtocol from '@app/protocol-validation/migrations/migrateProtocol';
 import validateProtocol from '@app/utils/validateProtocol';
 import { pruneProtocol } from '@app/utils/prune';
-import pruneProtocolAssets from '@app/utils/pruneProtocolAssets';
 import { archive, extract } from '@app/utils/protocols/lib/archive';
 import protocolTemplate from '@app/utils/protocolTemplate.json';
+import {
+  errors,
+  handleError,
+} from './errors';
+import {
+  getTempDir,
+  readProtocol,
+  writeProtocol,
+  deployNetcanvas,
+  commitNetcanvas,
+  revertNetcanvas,
+} from './lib';
 
-const errors = {
-  CreateFailed: 'CreateFailed', // Netcanvas file could not be generated
-  IncorrectPermissions: 'IncorrectPermissions', // File does not have read/write permissions
-  MigrationFailed: 'MigrationFailed', // Protocol could not be migrated
-  MissingSchemaVersion: 'MissingSchemaVersion', // Protocol does not contain schema version
-  NotFound: 'NotFound', // File could not be found
-  OpenFailed: 'OpenFailed', // Netcanvas file could not be opened
-  ReadError: 'ReadError', // File could not be read
-  SaveFailed: 'SaveFailed', // Netcanvas file could not be saved
-  VerificationFailed: 'VerificationFailed', // Netcanvas file could not be verifed
-  WriteError: 'WriteError', // File could not be written
-};
 
 const schemaVersionStates = {
   UPGRADE_APP: 'UPGRADE_APP',
@@ -33,85 +31,6 @@ const schemaVersionStates = {
 };
 
 const ProtocolsDidNotMatchError = new Error('Protocols did not match');
-
-/**
- * Helper function generator for use with `.catch()`. The original error
- * is logged, and then substituted for the custom error object, which is
- * thrown.
- * @param readableError An error object
- * @returns {function} A function that can be used inside .catch();
- */
-const getFriendlyError = (e, friendlyCode) => {
-  e.friendlyCode = friendlyCode;
-  return e;
-};
-
-const handleError = defaultError =>
-  (e) => {
-    log.error(e);
-
-    if (!e) {
-      throw getFriendlyError(new Error('No error to handle'), defaultError);
-    }
-
-    switch (e.code) {
-      case 'EACCES':
-        throw getFriendlyError(e, errors.IncorrectPermissions);
-      case 'ENOENT':
-        throw getFriendlyError(e, errors.NotFound);
-      default:
-        throw getFriendlyError(e, defaultError);
-    }
-  };
-
-/**
- * Essentially the same as path.join, but also creates the directory.
- * @returns {Promise} Resolves to path as a string
- */
-const getTempDir = (...args) => {
-  const dirPath = path.join(remote.app.getPath('temp'), 'architect', ...args);
-  return fse.mkdirp(dirPath)
-    .then(() => dirPath);
-};
-
-/**
- * Given the working path for a protocol (in /tmp/protocols `protocol.json`,
- * returns a promise that resolves to the parsed json object
- * @param {string} workingPath The protocol directory.
- * @returns {object} The protocol as an object
- */
-const readProtocol = (workingPath) => {
-  const protocolJsonPath = path.join(workingPath, 'protocol.json');
-
-  return fse.readJson(protocolJsonPath)
-    .catch(handleError(errors.ReadError));
-};
-
-/**
- * Given the working path for a protocol (in /tmp/protocols `protocol.json`.
- * Removes assets that aren't referenced in the protocol, and removes any
- * unsuported JSON values.
- * @param {string} workingPath The protocol directory.
- * @param {object} protocol the protocol data to write
- * @returns {Promise}
- */
-const writeProtocol = (workingPath, protocol) => {
-  const protocolJsonPath = path.join(workingPath, 'protocol.json');
-
-  const protocolWithDate = {
-    ...protocol,
-    lastModified: new Date().toISOString(),
-  };
-
-  return Promise.resolve()
-    .then(() => pruneProtocol(protocolWithDate))
-    .then(prunedProtocol =>
-      fse.writeJson(protocolJsonPath, prunedProtocol, { spaces: 2 })
-        .catch(handleError(errors.WriteError))
-        .then(() => pruneProtocolAssets(workingPath))
-        .then(() => prunedProtocol),
-    );
-};
 
 /**
  * @param {string} workingPath - working path in application /tmp/protocols/ dir
@@ -148,59 +67,6 @@ const importNetcanvas = filePath =>
         .then(() => destinationPath)
         .catch(handleError(errors.OpenFailed));
     });
-
-/**
- * Move a netcanvas file located in temporary directory into user space.
- * If the destination exists, make a backup copy of that file.
- *
- * @param netcanvasExportPath .netcanvas file path in temp
- * @param destinationUserPath Destination path
- * @returns {Promise} Resolves to { savePath, backupPath } if successful
- */
-const deployNetcanvas = (netcanvasExportPath, destinationUserPath) => {
-  const createBackup = true;
-  const f = path.parse(destinationUserPath);
-  const backupPath = path.join(f.dir, `${f.name}.backup-${new Date().getTime()}${f.ext}`);
-
-  return fse.pathExists(destinationUserPath)
-    .then((exists) => {
-      if (!exists || !createBackup) { return false; }
-
-      return fse.rename(destinationUserPath, backupPath)
-        .then(() => true);
-    })
-    .then(createdBackup =>
-      fse.rename(netcanvasExportPath, destinationUserPath)
-        .then(() => ({
-          savePath: destinationUserPath,
-          backupPath: createdBackup ? backupPath : null,
-        })),
-    );
-};
-
-const commitNetcanvas = ({ savePath, backupPath }) => {
-  if (!backupPath) { return Promise.resolve(savePath); }
-  // Check the new file definitely exists before deleting backup
-  return fse.stat(savePath)
-    .then((stat) => {
-      if (!stat.isFile()) { throw new Error('`savePath` does not exist'); }
-      return fse.unlink(backupPath)
-        .then(() => savePath);
-    });
-};
-
-const revertNetcanvas = ({ savePath, backupPath }) => {
-  if (!backupPath) { return Promise.resolve(savePath); } // Nothing to revert
-  // Check the backup definitely exists before deleting other file
-  return fse.stat(backupPath)
-    .then((stat) => {
-      if (!stat.isFile()) { throw new Error('`backupPath` does not exist'); }
-      return fse.unlink(savePath)
-        .then(() => fse.rename(backupPath, savePath))
-        .then(() => savePath);
-    });
-};
-
 /**
  * Create a new .netcanvas file at the target location.
  *
@@ -333,24 +199,14 @@ const migrateNetcanvas = (filePath, newFilePath, targetVersion = APP_SCHEMA_VERS
     )
     .catch(handleError(errors.MigrationFailed));
 
-const utils = {
-  commitNetcanvas,
-  createNetcanvasExport,
-  deployNetcanvas,
-  revertNetcanvas,
-  verifyNetcanvas,
-  writeProtocol,
-};
-
 export {
   checkSchemaVersion,
   createNetcanvas,
-  errors,
+  createNetcanvasExport,
+  verifyNetcanvas,
   importNetcanvas,
   migrateNetcanvas,
-  readProtocol,
   saveNetcanvas,
   schemaVersionStates,
-  utils,
   validateNetcanvas,
 };
